requirements.txt
pydeck>=0.8.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0

python-multipart>=0.0.6

plotly>=5.17.0
folium>=0.15.0
streamlit-folium>=0.15.0




